start abstraction of inputs and forms
