+++ packages/frontend-v2/components/Form/index.ts
@@ -1,0 +1,2 @@
+export { LoginForm } from "./LoginForm";
+export { SignUpForm } from "./SignUpForm";
