+++ packages/frontend-v2/components/index.ts
@@ -4,3 +4,4 @@ export * from "./Plan";
 export * from "./ScheduleCourse";
 export * from "./Spinner";
 export * from "./Button";
+export * from "./Form";
